+++ tests/typing/factory.py
@@ -55,13 +55,13 @@ animal7: Animal = provider7(1, 2, 3, b='1', c=2, e=0.0)
 provider8 = providers.FactoryDelegate(providers.Factory(object))
 
 # Test 9: to check FactoryAggregate provider
-provider9 = providers.FactoryAggregate(
-    a=providers.Factory(object),
-    b=providers.Factory(object),
+provider9: providers.FactoryAggregate[str] = providers.FactoryAggregate(
+    a=providers.Factory(str, "str1"),
+    b=providers.Factory(str, "str2"),
 )
-factory_a_9: providers.Factory = provider9.a
-factory_b_9: providers.Factory = provider9.b
-val9: Any = provider9('a')
+factory_a_9: providers.Factory[str] = provider9.a
+factory_b_9: providers.Factory[str] = provider9.b
+val9: str = provider9('a')
 
 # Test 10: to check the explicit typing
 factory10: providers.Provider[Animal] = providers.Factory(Cat)